Extract onAction to reusable hook
